+++ src/indicators/r028.rs
@@ -9,15 +9,15 @@ use crate::parse_pipe_separated_value;
 
 #[derive(Default)]
 pub struct R028 {
-    fixed_price_procurement_methods: HashSet<String>,
+    no_price_comparison_procurement_methods: HashSet<String>,
     price_comparison_procurement_methods: HashSet<String>,
 }
 
 impl Calculate for R028 {
     fn new(settings: &mut Settings) -> Self {
         Self {
-            fixed_price_procurement_methods: parse_pipe_separated_value(
-                settings.fixed_price_procurement_methods.clone(),
+            no_price_comparison_procurement_methods: parse_pipe_separated_value(
+                settings.no_price_comparison_procurement_methods.clone(),
             ),
             price_comparison_procurement_methods: parse_pipe_separated_value(
                 settings.price_comparison_procurement_methods.clone(),
@@ -26,8 +26,8 @@ impl Calculate for R028 {
     }
 
     fn fold(&self, item: &mut Indicators, release: &Map<String, Value>, ocid: &str) {
-        if Indicators::matches_procurement_method_details(release, &self.fixed_price_procurement_methods)
-            || Indicators::not_matches_procurement_method_details(release, &self.price_comparison_procurement_methods)
+        if Indicators::matches_procurement_method_details(release, &self.no_price_comparison_procurement_methods)
+            && Indicators::not_matches_procurement_method_details(release, &self.price_comparison_procurement_methods)
         {
             return;
         }