+++ src/indicators/util.rs
@@ -11,7 +11,7 @@ pub struct Tenderers {}
 
 #[derive(Default)]
 pub struct SecondLowestBidRatio {
-    fixed_price_procurement_methods: HashSet<String>,
+    no_price_comparison_procurement_methods: HashSet<String>,
     price_comparison_procurement_methods: HashSet<String>,
     currency: Option<String>,
 }
@@ -42,8 +42,8 @@ impl Calculate for Tenderers {
 impl Calculate for SecondLowestBidRatio {
     fn new(settings: &mut Settings) -> Self {
         Self {
-            fixed_price_procurement_methods: parse_pipe_separated_value(
-                settings.fixed_price_procurement_methods.clone(),
+            no_price_comparison_procurement_methods: parse_pipe_separated_value(
+                settings.no_price_comparison_procurement_methods.clone(),
             ),
             price_comparison_procurement_methods: parse_pipe_separated_value(
                 settings.price_comparison_procurement_methods.clone(),
@@ -53,8 +53,8 @@ impl Calculate for SecondLowestBidRatio {
     }
 
     fn fold(&self, item: &mut Indicators, release: &Map<String, Value>, ocid: &str) {
-        if Indicators::matches_procurement_method_details(release, &self.fixed_price_procurement_methods)
-            || Indicators::not_matches_procurement_method_details(release, &self.price_comparison_procurement_methods)
+        if Indicators::matches_procurement_method_details(release, &self.no_price_comparison_procurement_methods)
+            && Indicators::not_matches_procurement_method_details(release, &self.price_comparison_procurement_methods)
         {
             return;
         }