+++ docs/changelog.md
@@ -13,7 +13,7 @@
   - [R030](cli/indicators/R/030): (*Late bid won*).
   - [R048](cli/indicators/R/048): (*Heterogeneous supplier*).
   - [R058](cli/indicators/R/058): (*Heavily discounted bid*).
-  - `fixed_price_procurement_methods` configuration.
+  - `no_price_comparison_procurement_methods` configuration.
   - `price_comparison_procurement_methods` configuration.
 
 ### Changed